% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bggsoloranks.R
\docType{data}
\name{bggsoloranks}
\alias{bggsoloranks}
\title{2023 People's Choice Solo Game Rankings}
\format{
A data frame with 20699 observations of 4 variables.
\describe{
\item{\code{User}}{ BGG username}
\item{\code{Rank}}{ User's rank for a specific game. Users could provide 5-20 games in their ranking.}
\item{\code{Game}}{ Title of ranked game}
\item{\code{Score}}{ Scoring used in creating the top-200 list. (20-rank=score)}
}
}
\source{
Data collected by Kevin Erskine in 2023 to compile the One Player Guild's top 200 games in 2023.
\url{https://boardgamegeek.com/geeklist/324731/2023-peoples-choice-top-200-solo-games-1-200}
}
\usage{
data(bggsoloranks)
}
\description{
Rankings of singleplayer board games by BoardGameGeek users, collected by Kevin Erskine in 2023.
}
