% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\docType{data}
\name{gdpdata}
\alias{gdpdata}
\title{Historical GDP data for the 10 countries with the highest GDP in 2022.}
\format{
A data frame with 60 observations of 4 variables.
\describe{
\item{\code{country}}{ Country name}
\item{\code{code}}{ 3-letter code for the country}
\item{\code{year}}{ Year observed, every 5 years from 2000-2020, and 2022.}
\item{\code{gdp}}{ GDP}
}
}
\source{
Data collected by The World Bank, CC BY-4.0
\url{https://data.worldbank.org/indicator/NY.GDP.MKTP.CD}
}
\usage{
data(gdpdata)
}
\description{
Historical GDP data for the 10 countries with the highest GDP in 2022.
}
