% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sushi.R
\docType{data}
\name{sushi}
\alias{sushi}
\title{Sushi Ranking Preference Data (2003)}
\format{
A data frame with 4999 observations ranking 10 items.
\describe{
\item{\code{Suhsi index}}{0: ebi (shrimp)
1: anago (sea eel)
2: maguro (tuna)
3: ika (squid)
4: uni (sea urchin)
5: ikura (salmon roe)
6: tamago (egg)
7: toro (fatty tuna)
8: tekka maki (tuna roll)
9: kappa maki (cucumber roll)}
}
}
\source{
Data collected in T. Kamishima, "Nantonac Collaborative Filtering: Recommendation Based on Order Responses", KDD2003, pp.583-588 (2003)
}
\usage{
data(sushi)
}
\description{
Data on sushi preferences collected in
}
